Move DateFormatter to UI layer and improve Kover exclusions
1. Move DateFormatter from core.date to core.ui.date
   - DateFormatter contains @Composable function (rememberDateFormatter)
   - Belongs in UI layer, not core business logic
   - core.ui package is excluded from coverage

2. Improve Kover exclusions with universal patterns
   - Replace specific package "space.be1ski.vibits.shared.generated" with pattern "*.generated.*"
   - Replace specific package "space.be1ski.vibits.shared.core.ui" with pattern "*.ui.*"
   - More maintainable and future-proof

3. Remove DateFormatterTest
   - UI code doesn't need unit tests
   - @Composable functions cannot be tested in unit tests

This fixes coverage regression from PR #93 which moved DateFormatter from
platform package to testable core.date package, causing @Composable function
to appear as uncovered code.

Changes:
- Move DateFormatter.kt: core.date → core.ui.date
- Delete DateFormatterTest.kt
- Update imports in 10 view files
- Simplify Kover excludes to use universal patterns

Coverage: 94.51% (1792/1896)
