+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/app/view/TimeRangeControls.kt
@@ -27,9 +27,9 @@ import org.jetbrains.compose.resources.stringResource
 import space.be1ski.vibits.shared.app.domain.model.ActivityRange
 import space.be1ski.vibits.shared.app.domain.model.SuccessRateLevel
 import space.be1ski.vibits.shared.app.domain.usecase.GetSuccessRateLevelUseCase
-import space.be1ski.vibits.shared.core.date.DateFormatter
 import space.be1ski.vibits.shared.core.platform.date.currentLocalDate
 import space.be1ski.vibits.shared.core.ui.Indent
+import space.be1ski.vibits.shared.core.ui.date.DateFormatter
 import space.be1ski.vibits.shared.core.ui.theme.AppColors
 import space.be1ski.vibits.shared.core.ui.theme.resolve
 import space.be1ski.vibits.shared.feature.settings.domain.model.TimeRangeTab
