+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/memos/view/PostsScreen.kt
@@ -3,7 +3,7 @@ package space.be1ski.vibits.shared.feature.memos.view
 import androidx.compose.runtime.Composable
 import space.be1ski.vibits.shared.app.domain.model.ActivityMode
 import space.be1ski.vibits.shared.app.domain.model.ActivityRange
-import space.be1ski.vibits.shared.core.date.DateFormatter
+import space.be1ski.vibits.shared.core.ui.date.DateFormatter
 import space.be1ski.vibits.shared.feature.habits.domain.usecase.BuildActivityDataUseCase
 import space.be1ski.vibits.shared.feature.habits.domain.usecase.CalculateSuccessRateUseCase
 import space.be1ski.vibits.shared.feature.habits.view.StatsScreen
